// src/types/car.ts
export interface Car {
    id: string;
    brand: string;
    model: string;
    year: number;
    price: number;
    fuelType: string;
    mileage: number;
    transmission: string;
    seatingCapacity: number;
    color: string;
    description: string;
    features: string[];
    imageUrl: string;
    
  }